+++ forui/lib/widgets/time_field.dart
@@ -5,5 +5,7 @@
 /// See https://forui.dev/docs/form/time-field for working examples.
 library forui.widgets.time_field;
 
-export '../src/widgets/time_field/time_field.dart' hide FTimeFieldControllers;
+export '../src/widgets/time_field/time_field.dart';
+export '../src/widgets/time_field/time_field_controller.dart'
+    hide InternalFTimeFieldControl, InternalFTimeFieldController, Lifted, Managed;
 export '../src/widgets/time_field/time_field_style.dart';