+++ forui_internal_gen/lib/src/source/control_parent_mixin.dart
@@ -100,6 +100,10 @@ class ControlParentMixin {
 
   Method get defaultMethod => Method(
     (m) => m
+      ..docs.addAll([
+        '// TODO: https://github.com/dart-lang/sdk/issues/62198',
+        '// ignore: unused_element',
+      ])
       ..returns = refer(_returnType)
       ..name = '_default'
       ..requiredParameters.addAll([